final changes to the task api

+++ src/routers/task.js
@@ -1,23 +1,29 @@
 const express = require("express");
 const Task = require("../models/task.js");
 const router = new express.Router();
-const {sendtasks} = require("../emails/account.js");
+const auth = require('../middleware/auth.js');
 
-router.post('/tasks', async (req,res)=>{
-    const task = new Task(req.body)
+
+router.post('/tasks', auth,async (req,res)=>{
+
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    });
     try{
       const tasks = await task.save();
-      sendtasks(task.description);
+
       res.status(201).send(tasks); 
     }catch(e){
       res.status(400).send(e);
     }
 })
 
-router.get('/tasks',async (req,res)=>{
+router.get('/tasks',auth,async (req,res)=>{
     
     try{
-        const tasks =await Task.find({});
+        const tasks =await Task.find({owner: req.user._id});
+        // await req.user.populate('tasks').execPopulate();
         res.status(201).send(tasks);
     }catch(e){
         res.status(500).send();
@@ -25,11 +31,13 @@ router.get('/tasks',async (req,res)=>{
     
 })
 
-router.get('/tasks/:id',async (req,res)=>{
+router.get('/tasks/:id',auth,async (req,res)=>{
     const _id = req.params.id;
 
     try{
-        const task = await Task.findById(_id);
+        // const task = await Task.findById(_id);
+        const task = await Task.findOne({_id,owner: req.user._id});
+
         if(!task){
             res.status(401).send();
         }
@@ -39,7 +47,7 @@ router.get('/tasks/:id',async (req,res)=>{
     }
 })
 
-router.patch('/tasks/:id',async (req,res)=>{
+router.patch('/tasks/:id',auth,async (req,res)=>{
     const update = Object.keys(req.body);
     const allowedUpdates = ["description","completed"];
     const isValidOperation = update.every((update)=>{
@@ -51,7 +59,7 @@ router.patch('/tasks/:id',async (req,res)=>{
     }
     try{
 
-        const task = await Task.findByIdAndUpdate(req.params.id,req.body,{new:true,runValidators:true});
+        const task = await Task.findOne({_id: req.params.id, owner: req.user._id});
         if(!task){
             res.status(404).send();
         }
@@ -62,9 +70,10 @@ router.patch('/tasks/:id',async (req,res)=>{
 
 })
 
-router.delete('/tasks/:id',async (req,res)=>{
+router.delete('/tasks/:id',auth,async (req,res)=>{
     try{
-     const task = await Task.findByIdAndDelete(req.params.id);
+     const task = await Task.findByIdAndDelete({_id: req.params.id, owner: req.user._id});
+     
      if(!task){
          res.status(404).send();
      }
